Deduplicate init_cell_kdim_field_with_zero into common
The tmx copy was byte-identical to the dycore _wp stencil. Move it to
common/math/stencils under the dycore public names and merge both
docstrings; dycore and tmx now import the shared module.
