--- model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/stencils/init_cell_kdim_field_with_zero_wp.py
@@ -1,35 +1,0 @@
-# ICON4Py - ICON inspired code in Python and GT4Py
-#
-# Copyright (c) 2022-2024, ETH Zurich and MeteoSwiss
-# All rights reserved.
-#
-# Please, refer to the LICENSE file in the root directory.
-# SPDX-License-Identifier: BSD-3-Clause
-import gt4py.next as gtx
-from gt4py.next import broadcast
-
-from icon4py.model.common import dimension as dims, field_type_aliases as fa
-from icon4py.model.common.type_alias import wpfloat
-
-
-@gtx.field_operator
-def _init_cell_kdim_field_with_zero_wp() -> fa.CellKField[wpfloat]:
-    """Formerly known as _mo_solve_nonhydro_stencil_57 or _mo_solve_nonhydro_stencil_64."""
-    return broadcast(wpfloat("0.0"), (dims.CellDim, dims.KDim))
-
-
-@gtx.program(grid_type=gtx.GridType.UNSTRUCTURED)
-def init_cell_kdim_field_with_zero_wp(
-    field_with_zero_wp: fa.CellKField[wpfloat],
-    horizontal_start: gtx.int32,
-    horizontal_end: gtx.int32,
-    vertical_start: gtx.int32,
-    vertical_end: gtx.int32,
-) -> None:
-    _init_cell_kdim_field_with_zero_wp(
-        out=field_with_zero_wp,
-        domain={
-            dims.CellDim: (horizontal_start, horizontal_end),
-            dims.KDim: (vertical_start, vertical_end),
-        },
-    )
