--- model/atmosphere/subgrid_scale_physics/tmx/tests/tmx/stencil_tests/test_init_cell_kdim_field_with_zero.py
@@ -1,51 +1,0 @@
-# ICON4Py - ICON inspired code in Python and GT4Py
-#
-# Copyright (c) 2022-2024, ETH Zurich and MeteoSwiss
-# All rights reserved.
-#
-# Please, refer to the LICENSE file in the root directory.
-# SPDX-License-Identifier: BSD-3-Clause
-from typing import Any
-
-import gt4py.next as gtx
-import numpy as np
-import pytest
-
-from icon4py.model.atmosphere.subgrid_scale_physics.tmx.stencils.init_cell_kdim_field_with_zero import (
-    init_cell_kdim_field_with_zero,
-)
-from icon4py.model.common import dimension as dims
-from icon4py.model.common.grid import base
-from icon4py.model.common.type_alias import wpfloat
-from icon4py.model.common.utils import data_allocation as data_alloc
-from icon4py.model.testing.stencil_tests import StencilTest
-
-
-class TestInitCellKdimFieldWithZero(StencilTest):
-    PROGRAM = init_cell_kdim_field_with_zero
-    OUTPUTS = ("field",)
-
-    @staticmethod
-    def reference(
-        connectivities: dict[gtx.Dimension, np.ndarray],
-        *,
-        field: np.ndarray,
-        horizontal_start: int,
-        horizontal_end: int,
-        vertical_start: int,
-        vertical_end: int,
-        **kwargs: Any,
-    ) -> dict:
-        field_out = field.copy()
-        field_out[horizontal_start:horizontal_end, vertical_start:vertical_end] = 0.0
-        return dict(field=field_out)
-
-    @pytest.fixture
-    def input_data(self, grid: base.Grid) -> dict[str, Any]:
-        return dict(
-            field=data_alloc.random_field(grid, dims.CellDim, dims.KDim, dtype=wpfloat),
-            horizontal_start=0,
-            horizontal_end=gtx.int32(grid.num_cells),
-            vertical_start=0,
-            vertical_end=gtx.int32(grid.num_levels),
-        )
