+++ model/common/src/icon4py/model/common/math/stencils/init_cell_kdim_field_with_zero_wp.py
@@ -13,28 +13,28 @@ from icon4py.model.common.type_alias import wpfloat
 
 
 @gtx.field_operator
-def _init_cell_kdim_field_with_zero() -> fa.CellKField[wpfloat]:
+def _init_cell_kdim_field_with_zero_wp() -> fa.CellKField[wpfloat]:
     """
     Set a cell K field to zero.
 
-    Port of the ``CALL init(...)`` (mo_fortran_tools) zero fills of the tmx
-    scalar diffusion stages (mo_vdf.f90), e.g. the tracer and energy tendencies
-    at the top of 'Compute_diffusion_hydrometeors' /
-    'Compute_diffusion_temperature'.
+    Formerly known as _mo_solve_nonhydro_stencil_57 or
+    _mo_solve_nonhydro_stencil_64; also ports the ``CALL init(...)``
+    (mo_fortran_tools) zero fills, e.g. of the tracer and energy tendencies in
+    the tmx scalar diffusion stages (mo_vdf.f90).
     """
     return broadcast(wpfloat("0.0"), (dims.CellDim, dims.KDim))
 
 
 @gtx.program(grid_type=gtx.GridType.UNSTRUCTURED)
-def init_cell_kdim_field_with_zero(
-    field: fa.CellKField[wpfloat],
+def init_cell_kdim_field_with_zero_wp(
+    field_with_zero_wp: fa.CellKField[wpfloat],
     horizontal_start: gtx.int32,
     horizontal_end: gtx.int32,
     vertical_start: gtx.int32,
     vertical_end: gtx.int32,
 ) -> None:
-    _init_cell_kdim_field_with_zero(
-        out=field,
+    _init_cell_kdim_field_with_zero_wp(
+        out=field_with_zero_wp,
         domain={
             dims.CellDim: (horizontal_start, horizontal_end),
             dims.KDim: (vertical_start, vertical_end),
